Moved text from Runtime API to body row
The first line of the table is a heading by default. Since this table does not have headings, the run_inference function is in bold, which makes it look more important than the other rows and might be confusing. I moved it to the second row where it is equal to the other functions.