Add the -- dummy to ensure proper argument parsing
When running bash with -c, the first argument after the provided string
is treated as @0, *not* an argument to the script. When running the
existing command, the --no-install-kmd argument is ignored, and the
first argument would always be ignored. To fix this, let's add -- as a
dummy to become @0 so that all arguments are provided to the script as
intended.

Signed-off-by: June Knauth (tt) <jknauth@tenstorrent.com>